refactor/withdrawUser: 제미나이 코드리뷰관련 수정
