+++ src/main/java/org/example/eatopia/domain/user/repository/UserRepository.java
@@ -9,7 +9,6 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
 
     @Query("SELECT u FROM User u WHERE u.email = :email")
-        //이메일로 사용자를 조회하는 메소드
     Optional<User> findByEmail(String email);
 
 }
